Prueba.dicomshield/pyproject.toml
[project]
name = "dicomshield"
version = "0.1.0"
description = "Pseudoanonimizador DICOM configurable por YAML"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydicom>=2.4.0",
  "PyYAML>=6.0.0",
  "typer>=0.12.0",
]

[project.scripts]
dicomshield = "dicomshield.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
